Restructure src package
