+++ stories/Questions/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-import Readme from '../../quiz-us-components/src/QuestionMultipleChoice/README.md';
+import Readme from '../../quiz-us-components/src/components/QuestionMultipleChoice/README.md';
 import MultipleChoice from './MultipleChoice';
 
 export default () => {
